fix: move plugins to more logical place
