+++ packages/app-form-builder/src/admin/plugins/formsDataList/ImportButton/styledComponents.tsx
@@ -2,8 +2,8 @@ import styled from "@emotion/styled";
 import { Typography } from "@webiny/ui/Typography";
 import Accordion from "@webiny/app-page-builder/editor/plugins/elementSettings/components/Accordion";
 // assets
-import { ReactComponent as UploadInProgressIcon } from "../../../../icons/undraw-uploading.svg";
-import { ReactComponent as ExportInProgressIcon } from "../../../../icons/undraw_export_files.svg";
+import { ReactComponent as UploadInProgressIcon } from "../../../icons/undraw-uploading.svg";
+import { ReactComponent as ExportInProgressIcon } from "../../../icons/undraw_export_files.svg";
 import { ReactComponent as CircleIcon } from "@material-design-icons/svg/round/check_circle.svg";
 import { ReactComponent as CancelIcon } from "@material-design-icons/svg/round/cancel.svg";
